Cleaned up sources, almost ready for packaging
git-svn-id: file:///Users/avp/tmp/svn.20121212/LHPC/aff/trunk@114 65d2813f-f11f-0410-b17b-8647a28db151

+++ lib/tree-print.c
@@ -10,7 +10,7 @@ static void
 print_node(struct AffNode_s *node, void *arg)
 {
     printf(" %16p %016llx: [%16p %16p %16p %32s] ",
-	   node, node->id,
+	   node, (long long)node->id,
 	   node->children, node->next,
 	   node->key.parent, aff_symbol_name(node->key.name));
     switch (node->type) {
@@ -21,16 +21,16 @@ print_node(struct AffNode_s *node, void *arg)
 	printf("void\n");
 	break;
     case affNodeChar:
-	printf("char[%d] %016lld\n", node->size, node->offset);
+	printf("char[%d] %016lld\n", node->size, (long long)node->offset);
 	break;
     case affNodeInt:
-	printf("int[%d] %016lld\n", node->size, node->offset);
+	printf("int[%d] %016lld\n", node->size, (long long)node->offset);
 	break;
     case affNodeDouble:
-	printf("double[%d] %016lld\n", node->size, node->offset);
+	printf("double[%d] %016lld\n", node->size, (long long)node->offset);
 	break;
     case affNodeComplex:
-	printf("complex[%d] %016lld\n", node->size, node->offset);
+	printf("complex[%d] %016lld\n", node->size, (long long)node->offset);
 	break;
     default:
 	printf("unknown type %d\n", node->type);
@@ -46,7 +46,7 @@ aff_tree_print(struct AffTree_s *tt)
 	return;
     }
     printf("Tree table at %p, size=%lld, file-size=%lld:\n",
-	   tt, tt->size, tt->file_size);
+	   tt, (long long)tt->size, (long long)tt->file_size);
     aff_tree_foreach(tt, print_node, 0);
     printf("Tree table end\n");
 }
